+++ example/lib/screens/home_screen.dart
@@ -5,7 +5,6 @@ import 'package:flutter/material.dart';
 class HomeScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    print('building home scree');
     return Scaffold(
       appBar: AppBar(),
       body: Column(
@@ -17,15 +16,17 @@ class HomeScreen extends StatelessWidget {
                 child: Text("Second Screen"),
                 onPressed: () async {
                   ExtendedNavigator.ofRouter<Router>().pushNamed(
-                      Routes.secondScreen,
-                      arguments: SecondScreenArguments(message: 'title'),
-                      onReject: (guard) {
-                    showDialog(
+                    Routes.secondScreen,
+                    arguments: SecondScreenArguments(title: 'title'),
+                    onReject: (guard) {
+                      showDialog(
                         context: context,
                         builder: (_) => AlertDialog(
-                              title: Text('You need to be looged in'),
-                            ));
-                  });
+                          title: Text('You need to be logged in'),
+                        ),
+                      );
+                    },
+                  );
                 },
               ),
             ),