+++ example/lib/screens/second_screen.dart
@@ -1,12 +1,14 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:dio/dio.dart';
+import 'package:example/screens/login_screen.dart';
 import 'package:example/screens/nested_screens/nested_router.gr.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class SecondScreen extends StatelessWidget implements AutoRouteWrapper {
-  final String message;
 
-  SecondScreen({@required this.message});
+
+  SecondScreen(String title);
 
   final tabViews = <Widget>[
     Icon(Icons.book),